scripts/requirements.txt
google-generativeai>=0.8.0
feedparser>=6.0.0
python-slugify>=8.0.0
